Update weather icons and refactor js
The HTML code has been modified to give the weather icon an ID in the navigation bar, and the JavaScript file has been heavily refactored to include two new functions that retrieve weather icons based on the weather conditions. The XML response handling function has also been updated to use one of the new functions and to retrieve the weather group number.